requirements.txt
black
coveralls
pytest
pyusb==1.1.0
numpy==1.19.5
reprint==0.5.2
freezegun==1.0.0
pytest-cov
isort
flake8
websockets==8.1
jsonschema==3.2.0
gpsd-py3==0.3.0
async-timeout